separated schema and followed t3 structure
