--- packages/drizzle-db/index.ts
@@ -1,2 +1,0 @@
-export { neon } from "@neondatabase/serverless";
-export * from "drizzle-orm";
