+++ packages/drizzle-db/package.json
@@ -11,9 +11,11 @@
     "build": "echo 'Add build script here'",
     "test": "echo 'Add test script here'",
     "lint": "echo 'Add lint script here'",
-    "db:generate": "drizzle-kit generate:pg --schema=./src/drizzle/schema.ts --out=./migrations",
+    "db:generate": "drizzle-kit generate:pg --schema=./src/schema --out=./migrations",
+    "db:push": "pnpm with-env drizzle-kit push:pg",
     "db:migrate": "tsx ./src/drizzle/migrate.ts",
-    "db:studio": "pnpm drizzle-kit studio --port 3000 --verbose"
+    "db:studio": "pnpm with-env drizzle-kit studio --port 3000 --verbose",
+    "with-env": "dotenv -e ../../.env --"
   },
   "devDependencies": {
     "@types/pg": "^8.11.2",
