+++ packages/drizzle-db/src/schema/station.ts
@@ -1,0 +1,18 @@
+import { pgTable, text, timestamp } from "drizzle-orm/pg-core";
+
+import { restaurant } from "./restaurant";
+
+export const station = pgTable("Station", {
+  id: text("id").primaryKey().notNull(),
+  createdAt: timestamp("createdAt", { precision: 3, mode: "string" })
+    .defaultNow()
+    .notNull(),
+  updatedAt: timestamp("updatedAt", { precision: 3, mode: "string" }).notNull(),
+  name: text("name").notNull(),
+  restaurantId: text("restaurantId")
+    .notNull()
+    .references(() => restaurant.id, {
+      onDelete: "restrict",
+      onUpdate: "cascade",
+    }),
+});
