+++ packages/drizzle-db/src/schema/users.ts
@@ -1,0 +1,5 @@
+import { pgTable, text } from "drizzle-orm/pg-core";
+
+export const pushToken = pgTable("PushToken", {
+  token: text("token").primaryKey().notNull(),
+});
